Moved Electrode and Electrodes classes into submodules
This will make it easiser to merge in AMR/DAC_upgrades and use the improved GUI.
